feat: moved test to validate eip712 credentials to identity package
